+++ include/extras-kwidgetsaddons/libkactionselector.h
@@ -1797,7 +1797,7 @@ bool k_actionselector_is_full_screen(void* self);
 /// @param self KActionSelector*
 ///
 /// @return flag of enum Qt__WindowState
-int64_t k_actionselector_window_state(void* self);
+int32_t k_actionselector_window_state(void* self);
 
 /// Inherited from QWidget
 ///
@@ -1805,7 +1805,7 @@ int64_t k_actionselector_window_state(void* self);
 ///
 /// @param self KActionSelector*
 /// @param state flag of enum Qt__WindowState
-void k_actionselector_set_window_state(void* self, int64_t state);
+void k_actionselector_set_window_state(void* self, int32_t state);
 
 /// Inherited from QWidget
 ///
@@ -1813,7 +1813,7 @@ void k_actionselector_set_window_state(void* self, int64_t state);
 ///
 /// @param self KActionSelector*
 /// @param state flag of enum Qt__WindowState
-void k_actionselector_override_window_state(void* self, int64_t state);
+void k_actionselector_override_window_state(void* self, int32_t state);
 
 /// Inherited from QWidget
 ///
@@ -2330,7 +2330,7 @@ void k_actionselector_render3(void* self, void* target, void* targetOffset, void
 /// @param targetOffset QPoint*
 /// @param sourceRegion QRegion*
 /// @param renderFlags flag of enum QWidget__RenderFlag
-void k_actionselector_render4(void* self, void* target, void* targetOffset, void* sourceRegion, int64_t renderFlags);
+void k_actionselector_render4(void* self, void* target, void* targetOffset, void* sourceRegion, int32_t renderFlags);
 
 /// Inherited from QWidget
 ///
@@ -2360,7 +2360,7 @@ void k_actionselector_render32(void* self, void* painter, void* targetOffset, vo
 /// @param targetOffset QPoint*
 /// @param sourceRegion QRegion*
 /// @param renderFlags flag of enum QWidget__RenderFlag
-void k_actionselector_render42(void* self, void* painter, void* targetOffset, void* sourceRegion, int64_t renderFlags);
+void k_actionselector_render42(void* self, void* painter, void* targetOffset, void* sourceRegion, int32_t renderFlags);
 
 /// Inherited from QWidget
 ///
@@ -2377,7 +2377,7 @@ QPixmap* k_actionselector_grab1(void* self, void* rectangle);
 /// @param self KActionSelector*
 /// @param typeVal enum Qt__GestureType
 /// @param flags flag of enum Qt__GestureFlag
-void k_actionselector_grab_gesture2(void* self, int64_t typeVal, int64_t flags);
+void k_actionselector_grab_gesture2(void* self, int64_t typeVal, int32_t flags);
 
 /// Inherited from QWidget
 ///
